RA-1622 moved wage range exception out of validators folder.
