+++ src/Esfa.Vacancy.Application/Esfa.Vacancy.Application.csproj
@@ -72,6 +72,7 @@
     <Compile Include="Commands\CreateApprenticeship\IWageTypeMapper.cs" />
     <Compile Include="Commands\CreateApprenticeship\HourlyWageCalculator.cs" />
     <Compile Include="Commands\CreateApprenticeship\MinimumWageSelector.cs" />
+    <Compile Include="Commands\CreateApprenticeship\MissingWageRangeException.cs" />
     <Compile Include="Commands\CreateApprenticeship\Validators\ApplicationMethodValidator.cs" />
     <Compile Include="Commands\CreateApprenticeship\Validators\ContactDetailsValidator.cs" />
     <Compile Include="Commands\CreateApprenticeship\Validators\CandidateCriteriaValidator.cs" />
@@ -89,7 +90,6 @@
     <Compile Include="Commands\CreateApprenticeship\Validators\LocationTypeValidator.cs" />
     <Compile Include="Commands\CreateApprenticeship\Validators\LocationValidator.cs" />
     <Compile Include="Commands\CreateApprenticeship\Validators\LongDescriptionValidator.cs" />
-    <Compile Include="Commands\CreateApprenticeship\Validators\MissingWageRangeException.cs" />
     <Compile Include="Commands\CreateApprenticeship\Validators\NumberOfPositionsValidator.cs" />
     <Compile Include="Commands\CreateApprenticeship\Validators\ShortDescriptionValidator.cs" />
     <Compile Include="Commands\CreateApprenticeship\Validators\TitleValidator.cs" />
